docs(example): prep for moving items to server sdk
