+++ examples/authentication/regular-web-app/lib/middleware/hasteStrategy.js
@@ -1,0 +1,31 @@
+/* eslint-disable @typescript-eslint/no-unsafe-return */
+/* eslint-disable @typescript-eslint/no-unsafe-assignment */
+/* eslint-disable @typescript-eslint/no-unsafe-call */
+/* eslint-disable @typescript-eslint/no-unsafe-member-access */
+/* eslint-disable @typescript-eslint/no-unused-vars */
+import Auth0Strategy from 'passport-auth0';
+
+class HasteStrategy {
+  static initialize(options, verify) {
+    if (!options) {
+      options = {
+        domain: process.env.AUTH0_DOMAIN,
+        clientID: process.env.AUTH0_CLIENT_ID,
+        clientSecret: process.env.AUTH0_CLIENT_SECRET,
+        callbackURL: process.env.AUTH0_CALLBACK_URL || 'http://localhost:3002/',
+      };
+    }
+
+    if (!verify) {
+      verify = function (accessToken, refreshToken, extraParams, profile, done) {
+        // accessToken is the token to call Auth0 API (not needed in the most cases)
+        // extraParams.id_token has the JSON Web Token
+        // profile has all the information from the user
+        return done(null, profile);
+      };
+    }
+    return new Auth0Strategy(options, verify);
+  }
+}
+
+export default HasteStrategy;
