+++ examples/authentication/regular-web-app/public/stylesheets/style.css
@@ -58,6 +58,7 @@ a {
   transition: background-color 0.2s ease-in-out 0s, color 0.2s ease-in-out 0s;
   text-transform: none;
   letter-spacing: 0px;
+  line-height: 1;
 }
 
 .button:hover {
