+++ examples/authentication/regular-web-app/routes/hasteAuthRoutes.js
