+++ examples/authentication/regular-web-app/views/index.jade
@@ -2,11 +2,12 @@ extends layout
 
 block content
   h1= title
-  p Welcome to #{title}
   if locals.user
+    p Welcome #{user.displayName} to #{title}
     div(class="header")
       img(src=user.picture, class='logo')
       button(onclick="logout()" class="button") Sign Out
   else
+    p Welcome to #{title}
     button(onclick="login()" class="nobutton")
       img(src= './images/login.svg', class='logo')
